requirements.txt
flask==3.0.0
gspread==5.12.0
google-auth==2.23.0
google-auth-oauthlib==1.1.0
requests==2.31.0